Added aht10 drivers and updated controller to read temp,humidity as float.

+++ controller/core/event.go
@@ -35,8 +35,8 @@ type Motion struct {
 
 type Temp struct {
 	pktInfo
-	Temp     byte
-	Humidity byte
+	Temp     float32
+	Humidity float32
 }
 
 type Volt struct {
